Add support for defining a separate linker
Since sometimes a different linker program is used for linking the code
than for compiling it (i.e. such as in programming in assembly), one can
now specify a linker program such as 'ld'. If no linker is set
explicitly, this option defaults to the same program as the compiler.